--- src/components/pages/CustomPractice/customPractice.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const CustomPracticeDiv = styled.div``;
-
-export default function CustomPractice() {
-	return <CustomPracticeDiv>custom practice</CustomPracticeDiv>;
-}
