+++ src/components/pages/YourDecks/yourDecks.js
@@ -1,33 +1,13 @@
 import React from 'react';
-import styled from 'styled-components';
-import getFetch from '../../../utils/fetchData/get-fetch';
+import ViewDecks from '../../viewDecks/viewDecks';
 
-const YourDecksDiv = styled.div``;
-
-export default function YourDecks() {
-	const [decks, setDecks] = React.useState(null);
-	React.useEffect(() => {
-		const fetchObj = {
-			endpoint: 'decks',
-			errorMessage: 'could not get your decks',
-			authRequired: true,
-		};
-		getFetch(fetchObj).then((res) => {
-			console.log(res);
-			setDecks(res);
-			// console.log(res, typeof res)
-			// // setDecks(JSON.parse(res))
-		});
-	}, []);
+export default function MyDecks() {
+	const [myState, setMyState] = React.useState(null);
 
 	return (
-		<YourDecksDiv>
-			<h2>Your Decks</h2>
-			{decks ? (
-				decks.map((deck) => <p>{deck.deck_name}</p>)
-			) : (
-				<h3>This is where your decks will appear...</h3>
-			)}
-		</YourDecksDiv>
+		<ViewDecks
+		// selectedDecks={selectedDecks}
+		// setSelectedDecks={setSelectedDecks}
+		/>
 	);
 }
